Refactor around noise and micro optimise (#197)

+++ steel-worldgen/src/biomes/biome_source.rs
@@ -395,12 +395,12 @@ impl EndChunkBiomeSampler<'_> {
         {
             return erosion;
         }
-        let weird_block_x = (chunk_x * 2 + 1) * 8;
-        let weird_block_z = (chunk_z * 2 + 1) * 8;
+        let weird_block_x = f64::from((chunk_x * 2 + 1) * 8);
+        let weird_block_z = f64::from((chunk_z * 2 + 1) * 8);
         let erosion = self
             .source
             .end_islands
-            .sample(weird_block_x, 0, weird_block_z);
+            .sample(weird_block_x, 0.0, weird_block_z);
         self.cached_erosion = Some((chunk_x, chunk_z, erosion));
         erosion
     }
